IssueNo:#IANXDU Description:修改资源Id类型为uint32_t挑单 Sig:bundleManager Feature or Bugfix:BugFix Binary Source:No

+++ services/bundlemgr/src/bundle_resource/bundle_resource_parser.cpp
@@ -211,7 +211,7 @@ bool BundleResourceParser::ParseResourceInfoWithSameHap(const int32_t userId, Re
 }
 
 bool BundleResourceParser::ParseLabelResourceByPath(
-    const std::string &hapPath, const int32_t labelId, std::string &label)
+    const std::string &hapPath, const uint32_t labelId, std::string &label)
 {
     if (hapPath.empty()) {
         APP_LOGE("hapPath is empty");
@@ -238,7 +238,7 @@ bool BundleResourceParser::ParseLabelResourceByPath(
     return true;
 }
 
-bool BundleResourceParser::ParseIconResourceByPath(const std::string &hapPath, const int32_t iconId,
+bool BundleResourceParser::ParseIconResourceByPath(const std::string &hapPath, const uint32_t iconId,
     ResourceInfo &resourceInfo)
 {
     if (hapPath.empty()) {
@@ -287,7 +287,7 @@ bool BundleResourceParser::ParseResourceInfoByResourceManager(
 
 bool BundleResourceParser::ParseLabelResourceByResourceManager(
     const std::shared_ptr<Global::Resource::ResourceManager> resourceManager,
-    const int32_t labelId, std::string &label)
+    const uint32_t labelId, std::string &label)
 {
     if (resourceManager == nullptr) {
         APP_LOGE("resourceManager is nullptr");
@@ -297,7 +297,7 @@ bool BundleResourceParser::ParseLabelResourceByResourceManager(
         APP_LOGW("ParseLabelResource labelId is 0 or less than 0, label is bundleName");
         return false;
     }
-    auto ret = resourceManager->GetStringById(static_cast<uint32_t>(labelId), label);
+    auto ret = resourceManager->GetStringById(labelId, label);
     if (ret != OHOS::Global::Resource::RState::SUCCESS) {
         APP_LOGE("GetStringById failed %{public}d, labelId %{public}d",
             static_cast<int32_t>(ret), labelId);
